# api-contracts/backend/openapi.json
{"openapi":"3.1.0","info":{"title":"MKopp.Dev Backend API","description":"RESTful API for MKopp.Dev personal website and blog platform","contact":{"name":"Michael Kopp","url":"https://mkopp.dev","email":"michaeljg.kopp@gmail.com"},"license":{"name":"MIT License","url":"https://github.com/MichaelJGKopp/mkopp.dev/blob/main/LICENSE"},"version":"1.0.0"},"servers":[{"url":"https://api.mkopp.dev","description":"Docker Production Server"},{"url":"http://api.localhost","description":"Docker Development Server"},{"url":"http://localhost:8200","description":"Local Development Server"}],"tags":[{"name":"Likes","description":"Blog like management API"},{"name":"User","description":"User management API"},{"name":"Blog","description":"Blog post management API"},{"name":"Comments","description":"Blog comment management API"}],"paths":{"/v1/blog/{id}":{"put":{"tags":["Blog"],"summary":"Update blog post","operationId":"updatePost","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BlogPostRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BlogPostResponse"}}}}},"security":[{"oauth2":[]}]},"delete":{"tags":["Blog"],"summary":"Delete blog post","operationId":"deletePost","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK"}},"security":[{"oauth2":[]}]}},"/v1/comments/{commentId}/like":{"get":{"tags":["Comments"],"summary":"Get comment like count and user like status","operationId":"getCommentLikeInfo","parameters":[{"name":"commentId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CommentLikeResponse"}}}}}},"post":{"tags":["Comments"],"summary":"Toggle like on comment","operationId":"toggleCommentLike","parameters":[{"name":"commentId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CommentLikeResponse"}}}}},"security":[{"oauth2":[]}]}},"/v1/blog":{"get":{"tags":["Blog"],"summary":"Get all blog posts","operationId":"getAllPosts","parameters":[{"name":"pageable","in":"query","required":true,"schema":{"$ref":"#/components/schemas/Pageable"}}],"responses":{"200":{"description":"Successfully retrieved blog posts","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PageBlogPostResponse"}}}}}},"post":{"tags":["Blog"],"summary":"Create blog post","operationId":"createPost","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BlogPostRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BlogPostResponse"}}}}},"security":[{"oauth2":[]}]}},"/v1/blog/{blogPostId}/like":{"get":{"tags":["Likes"],"summary":"Get like info","operationId":"getLikeInfo","parameters":[{"name":"blogPostId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/LikeResponse"}}}}}},"post":{"tags":["Likes"],"summary":"Toggle like","operationId":"toggleLike","parameters":[{"name":"blogPostId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/LikeResponse"}}}}},"security":[{"oauth2":[]}]}},"/v1/blog/{blogPostId}/comments":{"get":{"tags":["Comments"],"summary":"Get top-level comments for blog post","operationId":"getTopLevelComments","parameters":[{"name":"blogPostId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}},{"name":"pageable","in":"query","required":true,"schema":{"$ref":"#/components/schemas/Pageable"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PageCommentTreeItem"}}}}}},"post":{"tags":["Comments"],"summary":"Create comment","operationId":"createComment","parameters":[{"name":"blogPostId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CommentRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CommentTreeItem"}}}}},"security":[{"oauth2":[]}]}},"/v1/comments/{commentId}":{"delete":{"tags":["Comments"],"summary":"Delete comment","operationId":"deleteComment","parameters":[{"name":"commentId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK"}},"security":[{"oauth2":[]}]},"patch":{"tags":["Comments"],"summary":"Update comment (PATCH)","operationId":"updateComment","parameters":[{"name":"commentId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CommentRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CommentTreeItem"}}}}},"security":[{"oauth2":[]}]}},"/v1/users/{userId}":{"get":{"tags":["User"],"summary":"Get user by ID","operationId":"getUserById","parameters":[{"name":"userId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}}}}},"/v1/users/me":{"get":{"tags":["User"],"summary":"Get current user","operationId":"getCurrentUser","responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}}},"security":[{"oauth2":[]}]}},"/v1/comments/{commentId}/replies":{"get":{"tags":["Comments"],"summary":"Get replies for a comment","operationId":"getReplies","parameters":[{"name":"commentId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}},{"name":"pageable","in":"query","required":true,"schema":{"$ref":"#/components/schemas/Pageable"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PageCommentTreeItem"}}}}}}},"/v1/blog/{slug}":{"get":{"tags":["Blog"],"summary":"Get blog post by slug","operationId":"getPostBySlug","parameters":[{"name":"slug","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Successfully retrieved blog post","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BlogPostResponse"}}}},"404":{"description":"Blog post not found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BlogPostResponse"}}}}}}},"/v1/blog/{blogPostId}/comments/count":{"get":{"tags":["Comments"],"summary":"Get comment count for blog post","operationId":"getCommentCount","parameters":[{"name":"blogPostId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"integer","format":"int64"}}}}}}},"/v1/blog/tag/{tagName}":{"get":{"tags":["Blog"],"summary":"Get blog posts by tag","operationId":"getPostsByTag","parameters":[{"name":"tagName","in":"path","required":true,"schema":{"type":"string"}},{"name":"pageable","in":"query","required":true,"schema":{"$ref":"#/components/schemas/Pageable"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PageBlogPostResponse"}}}}}}}},"components":{"schemas":{"BlogPostRequest":{"type":"object","description":"Blog post creation/update request","properties":{"slug":{"type":"string","description":"Unique slug","example":"my-first-post"},"title":{"type":"string","description":"Title"},"description":{"type":"string","description":"Description"},"content":{"type":"string","description":"Content"},"publishedAt":{"type":"string","format":"date-time","description":"Publication date"},"thumbnailUrl":{"type":"string","description":"Thumbnail URL"},"type":{"type":"string","description":"Post type","enum":["BLOG","YOUTUBE","EXTERNAL"]},"externalUrl":{"type":"string","description":"External URL"},"tags":{"type":"array","description":"Tags","items":{"type":"string"},"uniqueItems":true}}},"BlogPostResponse":{"type":"object","description":"Blog post response","properties":{"id":{"type":"string","format":"uuid","description":"ID"},"slug":{"type":"string","description":"Slug"},"title":{"type":"string","description":"Title"},"description":{"type":"string","description":"Description"},"content":{"type":"string","description":"Content"},"authorId":{"type":"string","format":"uuid","description":"Author ID"},"authorName":{"type":"string","description":"Author name"},"publishedAt":{"type":"string","format":"date-time","description":"Publication date"},"createdAt":{"type":"string","format":"date-time","description":"Creation date"},"updatedAt":{"type":"string","format":"date-time","description":"Update date"},"thumbnailUrl":{"type":"string","description":"Thumbnail URL"},"type":{"type":"string","description":"Post type","enum":["BLOG","YOUTUBE","EXTERNAL"]},"externalUrl":{"type":"string","description":"External URL"},"tags":{"type":"array","description":"Tags","items":{"type":"string"},"uniqueItems":true}}},"CommentLikeResponse":{"type":"object","properties":{"count":{"type":"integer","format":"int64"},"isLiked":{"type":"boolean"}}},"LikeResponse":{"type":"object","description":"Like information","properties":{"likeCount":{"type":"integer","format":"int64","description":"Like count"},"isLiked":{"type":"boolean","description":"Is liked by user"}}},"CommentRequest":{"type":"object","description":"Comment creation/update request","properties":{"content":{"type":"string","description":"Content"},"parentCommentId":{"type":"string","format":"uuid","description":"Parent comment ID"}}},"CommentTreeItem":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"userId":{"type":"string","format":"uuid"},"content":{"type":"string"},"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"},"replyCount":{"type":"integer","format":"int64"},"likeCount":{"type":"integer","format":"int64"}}},"UserResponse":{"type":"object","description":"User response","properties":{"id":{"type":"string","format":"uuid","description":"User ID"},"username":{"type":"string","description":"Username"},"email":{"type":"string","description":"Email"},"firstName":{"type":"string","description":"First name"},"lastName":{"type":"string","description":"Last name"},"roles":{"type":"array","description":"Roles","items":{"type":"string"}},"tokenExpiry":{"type":"string","format":"date-time","description":"Token expiry"}}},"Pageable":{"type":"object","properties":{"page":{"type":"integer","format":"int32","minimum":0},"size":{"type":"integer","format":"int32","minimum":1},"sort":{"type":"array","items":{"type":"string"}}}},"PageCommentTreeItem":{"type":"object","properties":{"totalPages":{"type":"integer","format":"int32"},"totalElements":{"type":"integer","format":"int64"},"size":{"type":"integer","format":"int32"},"content":{"type":"array","items":{"$ref":"#/components/schemas/CommentTreeItem"}},"number":{"type":"integer","format":"int32"},"sort":{"$ref":"#/components/schemas/SortObject"},"first":{"type":"boolean"},"last":{"type":"boolean"},"numberOfElements":{"type":"integer","format":"int32"},"pageable":{"$ref":"#/components/schemas/PageableObject"},"empty":{"type":"boolean"}}},"PageableObject":{"type":"object","properties":{"offset":{"type":"integer","format":"int64"},"sort":{"$ref":"#/components/schemas/SortObject"},"pageNumber":{"type":"integer","format":"int32"},"paged":{"type":"boolean"},"pageSize":{"type":"integer","format":"int32"},"unpaged":{"type":"boolean"}}},"SortObject":{"type":"object","properties":{"empty":{"type":"boolean"},"unsorted":{"type":"boolean"},"sorted":{"type":"boolean"}}},"PageBlogPostResponse":{"type":"object","properties":{"totalPages":{"type":"integer","format":"int32"},"totalElements":{"type":"integer","format":"int64"},"size":{"type":"integer","format":"int32"},"content":{"type":"array","items":{"$ref":"#/components/schemas/BlogPostResponse"}},"number":{"type":"integer","format":"int32"},"sort":{"$ref":"#/components/schemas/SortObject"},"first":{"type":"boolean"},"last":{"type":"boolean"},"numberOfElements":{"type":"integer","format":"int32"},"pageable":{"$ref":"#/components/schemas/PageableObject"},"empty":{"type":"boolean"}}}},"securitySchemes":{"bearer-jwt":{"type":"http","description":"Enter your JWT token obtained from Keycloak","scheme":"bearer","bearerFormat":"JWT"},"keycloak-oauth2":{"type":"oauth2","description":"OAuth2 authentication with Keycloak (Authorization Code Flow)","flows":{"authorizationCode":{"authorizationUrl":"http://auth.localhost/realms/mkopp/protocol/openid-connect/auth","tokenUrl":"http://auth.localhost/realms/mkopp/protocol/openid-connect/token","scopes":{"openid":"OpenID Connect scope","profile":"User profile information","email":"User email address"}}}}}}}