# libs/api-clients/backend/model/sortObject.ts
/**
 * MKopp.Dev Backend API
 *
 * Contact: michaeljg.kopp@gmail.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface SortObject { 
    empty?: boolean;
    unsorted?: boolean;
    sorted?: boolean;
}