# libs/api-clients/backend/model/userResponse.ts
/**
 * MKopp.Dev Backend API
 *
 * Contact: michaeljg.kopp@gmail.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * User response
 */
export interface UserResponse { 
    /**
     * User ID
     */
    id?: string;
    /**
     * Username
     */
    username?: string;
    /**
     * Email
     */
    email?: string;
    /**
     * First name
     */
    firstName?: string;
    /**
     * Last name
     */
    lastName?: string;
    /**
     * Roles
     */
    roles?: Array<string>;
    /**
     * Token expiry
     */
    tokenExpiry?: string;
}